update qosmanager datamodel and add qoscalculator model

+++ it.unipi.iotplatform.qosbroker.qosmanager/src/main/java/it/unipi/iotplatform/qosbroker/qosmanager/datamodel/QoSreq.java
@@ -24,31 +24,31 @@ public class QoSreq extends DataStructure {
 	
 	@XmlElement(name = "maxResponseTime")
 	@JsonProperty("maxResponseTime")
-	private Integer maxResponseTime;
+	private Double maxResponseTime;
 	
 	@XmlElement(name = "maxRateRequest")
 	@JsonProperty("maxRateRequest")
-	private Integer maxRateRequest;
+	private Double maxRateRequest;
 
 	public QoSreq() {
 
-		this.maxResponseTime = 0;
-		this.maxRateRequest = 0;
+		this.maxResponseTime = 0.0;
+		this.maxRateRequest = 0.0;
 	}
 
-	public Integer getMaxResponseTime() {
+	public Double getMaxResponseTime() {
 		return maxResponseTime;
 	}
 
-	public void setMaxResponseTime(Integer maxResponseTime) {
+	public void setMaxResponseTime(Double maxResponseTime) {
 		this.maxResponseTime = maxResponseTime;
 	}
 
-	public Integer getMaxRateRequest() {
+	public Double getMaxRateRequest() {
 		return maxRateRequest;
 	}
 
-	public void setMaxRateRequest(Integer maxRateRequest) {
+	public void setMaxRateRequest(Double maxRateRequest) {
 		this.maxRateRequest = maxRateRequest;
 	}
